pyproject.toml
[tool.poetry]
name = "blogproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.6"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
pre-commit = "^3.4.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
django-stubs = "^4.2.4"
djangorestframework-stubs = "^3.14.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length=120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
# Same line length as black formatter to prevent conflicts
line_length = 120
skip = ["migrations"]